Correct usage of the apostrophe parser (solve #29)
We had a stupid apostrophe parsing problem because we were parsing the
whole text instead of doing that line by line as the parser expects.
I added tests for this weird case.

Also: update apostrophe.py to be closer to MediaWiki's code.

Also: add docstring for the parameters of html.make_parser()