Make <body> tag generation optional (closes #31).
Parsing with the 'wikitext' parser will parse and render as expected and
do nothing more. This is useful for the inclusion of the article content in
a larger HTML page.

Parsing with the 'body' parser will also add the <body>...</body> tags
and the interwiki links and category links. This is useful to mimic the
behavior of MediaWiki's parser when rendering a whole article as a web
page.